fix mua 1 service + return user + so vote

+++ src/main/java/com/feature/gcoin/service/VoteService.java
@@ -17,7 +17,7 @@ public interface VoteService {
     // choose best staff
     boolean addBestStaff(String address);
 
-    Long getNumberOfVote(String address);
+    int getNumberOfVote(String address);
 
     boolean voteToStaff(VoteDTO voteDTO);
